+++ resources/config/config.php
@@ -22,4 +22,42 @@ return [
       |
       */
       'ttl' => env('BUILDER_TTL', 60),
+
+      /*
+      |--------------------------------------------------------------------------
+      | Public path
+      |--------------------------------------------------------------------------
+      |
+      | Where templates are stored in the public folder
+      |
+      | i.e. 'app/storage/streams/default/builder'
+      |
+      |
+      */
+      'path' => env('BUILDER_PATH', 'builder'),
+
+      /*
+      |--------------------------------------------------------------------------
+      | Builder template archive url
+      |--------------------------------------------------------------------------
+      |
+      | Url template used to download a template from its repo
+      |
+      |
+      */
+      'archive' => env('BUILDER_ARCHIVE_URL', 'http://github.com/{{ registry }}/{{ template }}/archive/master.zip'),
+
+      /*
+      |--------------------------------------------------------------------------
+      | Builder template temp file
+      |--------------------------------------------------------------------------
+      |
+      | Location and name of template compressed file which is stored in the
+      | builder path (see above)
+      |
+      |
+      */
+      'tmp' => env('BUILDER_TMP', 'tmp/master.zip'),
+
+
 ];